fix move keypair
